Refactor 078.
